Warn when trying to export with missing public path

+++ packages/react-cosmos/src/server/export.js
@@ -56,7 +56,14 @@ export default function startExport() {
   if (publicPath) {
     if (outputPath.indexOf(publicPath) === -1) {
       const exportPublicPath = path.join(outputPath, publicUrl);
-      fs.copySync(publicPath, exportPublicPath);
+      if (fs.existsSync(publicPath)) {
+        fs.copySync(publicPath, exportPublicPath);
+      } else {
+        console.log(
+          '[Cosmos] Warning: config.publicPath points to missing dir',
+          publicPath
+        );
+      }
     } else {
       console.warn(
         `[Cosmos] Warning: Can't export public path because it contains the export path! (avoiding infinite loop)`
